% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admissible_doses.R
\name{admissible_doses}
\alias{admissible_doses}
\title{Find doses which would not be 'too big' an increase}
\usage{
admissible_doses(current_dose, doses)
}
\arguments{
\item{current_dose}{Row number of 'doses' corresponding ot current dose.}

\item{doses}{A matrix with 2 columns, where each row is a unique dose combination (on original scale)}
}
\description{
Find doses which would not be 'too big' an increase
}

