% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom_2_bern.R
\name{binom_2_bern}
\alias{binom_2_bern}
\title{Turn a short DLT data frame with n_dlt/n into a longer version}
\usage{
binom_2_bern(df)
}
\arguments{
\item{df}{A data frame with columns 'dose_1', 'dose_2', 'n' and 'n_dlt'}
}
\description{
Convert a data frame with binomial data into a longer data frame of binary responses.
This is useful to give to JAGS.
}

