% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_prior.R
\name{choose_prior}
\alias{choose_prior}
\title{Find prior parameters for an increasing logistic curve}
\usage{
choose_prior(drug_1_dose_1, drug_1_dose_1_best_guess, drug_1_dose_2,
  drug_1_dose_2_best_guess, drug_1_n_prior, drug_2_dose_1,
  drug_2_dose_1_best_guess, drug_2_dose_2, drug_2_dose_2_best_guess,
  drug_2_n_prior, interaction_sd)
}
\arguments{
\item{drug_1_dose_1}{A dose level for drug 1. On original scale.}

\item{drug_1_dose_1_best_guess}{The best guess for p(DLT) at 'dose_1'.}

\item{drug_1_dose_2}{A dose level for drug 2. On original scale. Must be the same as 'd_star[1]'.}

\item{drug_1_dose_2_best_guess}{The best guess for p(DLT) at 'dose_2'.}

\item{drug_1_n_prior}{The number of prior observations that is roughly equivalent to strength of belief.}

\item{drug_2_dose_1}{A dose level for drug 1. On original scale.}

\item{drug_2_dose_1_best_guess}{The best guess for p(DLT) at 'dose_1'.}

\item{drug_2_dose_2}{A dose level for drug 2. On original scale. Must be the same as 'd_star[2]'.}

\item{drug_2_dose_2_best_guess}{The best guess for p(DLT) at 'dose_2'.}

\item{drug_2_n_prior}{The number of prior observations that is roughly equivalent to strength of belief.}

\item{interaction_sd}{The standard deviation of the interaction parameter.}
}
\value{
A vector of prior parameters.
}
\description{
Find prior parameters for an increasing logistic curve
}

