% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_doses.R
\name{combine_doses}
\alias{combine_doses}
\title{Choose dose combinations to be investigated, and plot the prior probabilities of toxicity.}
\usage{
combine_doses(mono_model_1, mono_model_2, eta_mean, eta_prec, doses_1, doses_2,
  over_limit)
}
\arguments{
\item{mono_model_1}{An object of class "LogisticLogNormal".}

\item{mono_model_2}{An object of class "LogisticLogNormal".}

\item{eta_mean}{Prior mean for interaction parameter.}

\item{eta_prec}{Prior precision for interaction parameter.}

\item{doses_1}{The doses of drug 1 to be tested.}

\item{doses_2}{The doses of drug 2 to be tested.}

\item{over_limit}{A number between 0 and 1. If the probability of overdosing is more than 'over_limit'
then this dose will be declared inadmissible.}
}
\value{
A list of 8 items:
- dose combinations on original scale,
- dose combinations on log scale,
- reference dose 1,
- reference dose 2,
- over_limit (maximum acceptable probability of overdosing)
- plot of prior distribution,
- plot of prior distribution including marginals.
}
\description{
Choose dose combinations to be investigated, and plot the prior probabilities of toxicity.
}

