% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_doses_trim.R
\name{combine_doses_trim}
\alias{combine_doses_trim}
\title{Choose dose combinations to be investigated, and plot the prior probabilities of toxicity.}
\usage{
combine_doses_trim(combo_doses, drop_ids)
}
\arguments{
\item{combo_doses}{A list from "combine_doses" function.}

\item{drop_ids}{Which elements of combo_doses$orig (and combo_doses$log_scale) should be dropped.}
}
\value{
A list of 8 items:
- dose combinations on original scale,
- dose combinations on log scale,
- reference dose 1,
- reference dose 2,
- over_limit (maximum acceptable probability of overdosing)
- plot of prior distribution,
- plot of prior distribution including marginals.
}
\description{
Choose dose combinations to be investigated, and plot the prior probabilities of toxicity.
}

