% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_all.R
\name{compare_all}
\alias{compare_all}
\title{Graphical comparison of a model-based design versus a rule-based design}
\usage{
compare_all(model_sum, rule_sum)
}
\arguments{
\item{model_sim}{a summary of the model based simulation -- from 'sum_sims'.}

\item{rule_sim}{a summary of the model based simulation -- an object from 'crmPack'.}
}
\value{
A ggplot object.
}
\description{
Graphical comparison of a model-based design versus a rule-based design
}

