% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_over_target.R
\name{compare_over_target}
\alias{compare_over_target}
\title{Plot the number of patients treated above the 'target' toxicity level}
\usage{
compare_over_target(model_sum, rule_sum)
}
\arguments{
\item{model_sim}{a summary of the model based simulation -- from 'sum_sims'.}

\item{rule_sim}{a summary of the model based simulation -- an object from 'crmPack'.}
}
\value{
A ggplot object.
}
\description{
Plot the number of patients treated above the 'target' toxicity level
}

