% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_overall_n.R
\name{compare_overall_n}
\alias{compare_overall_n}
\title{Plot the sample sizes of a series of simulated trials.}
\usage{
compare_overall_n(model_sum, rule_sum)
}
\arguments{
\item{model_sim}{a summary of the model based simulation -- from 'sum_sims'.}

\item{rule_sim}{a summary of the model based simulation -- an object from 'crmPack'.}
}
\value{
A ggplot object.
}
\description{
Plot the sample sizes of a series of simulated trials.
}

