% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_prop_dlts.R
\name{compare_prop_dlts}
\alias{compare_prop_dlts}
\title{Plot the proportion of DLTs in a series of simulated trials.}
\usage{
compare_prop_dlts(model_sum, rule_sum)
}
\arguments{
\item{model_sim}{a summary of the model based simulation -- from 'sum_sims'.}

\item{rule_sim}{a summary of the model based simulation -- an object from 'crmPack'.}
}
\value{
A ggplot object.
}
\description{
Plot the proportion of DLTs in a series of simulated trials.
}

