% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_true_tox.R
\name{compare_true_tox}
\alias{compare_true_tox}
\title{Plot the true toxicity probabilities at the selected doses}
\usage{
compare_true_tox(model_sum, rule_sum)
}
\arguments{
\item{model_sim}{a summary of the model based simulation -- from 'sum_sims'.}

\item{rule_sim}{a summary of the model based simulation -- an object from 'crmPack'.}
}
\value{
A ggplot object.
}
\description{
Plot the true toxicity probabilities at the selected doses
}

