% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_sim.R
\name{comparison_sim}
\alias{comparison_sim}
\title{Simulate model-based and rule-based trials, and compare results.}
\usage{
comparison_sim(truth, doses, starting_dose, route_through_doses, max_n_cohort,
  cohort_size, n_sims, seed)
}
\arguments{
\item{truth}{A vector of 'true' p(DLT)'s, corresponding to the rows of 'doses'.}

\item{doses}{A list containing doses on the original and log scale.}

\item{starting_dose}{The row number of 'doses' corresponding to the starting dose.}

\item{route_through_doses}{The row numbers of 'doses' corresponding to the doses used in the '3+3' design.}

\item{max_n_cohort}{The maximum number of cohorts.}

\item{cohort_size}{The number of patients per cohort.}

\item{n_sims}{The number of simulations.}

\item{seed}{The random number seed.}
}
\value{
A list of 4 items: a plot of the true p(DLTs), a graphical comparison of the 2 methods, a summary of model-based simulations, and a summary of rule-based simulations.
}
\description{
Simulate model-based and rule-based trials, and compare results.
}

