% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_combo_prior.R
\name{create_combo_prior}
\alias{create_combo_prior}
\title{Create a prior for combination dose response model from 2 crmPack model objects}
\usage{
create_combo_prior(model_1, model_2, eta_mean = 0, eta_prec = 0.5)
}
\arguments{
\item{model_1}{First crmPack model: prior for drug 1.}

\item{model_2}{Second crmPack model: prior for drug 2.}

\item{eta_mean}{Prior mean for the interaction parameter}

\item{eta_prec}{Prior precision for the interaction parameter.}
}
\value{
A list of prior means/variances for parameters in the combination dose-response model.
}
\description{
Create a prior for combination dose response model from 2 crmPack model objects
}

