% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_row.R
\name{dose_row}
\alias{dose_row}
\title{Find the row number of a dose combination of interest.}
\usage{
dose_row(doses, dose, orig_scale = FALSE)
}
\arguments{
\item{doses}{A list containing doses on the original and log scale.}

\item{dose}{The dose combination of interest.}

\item{orig_scale}{A logical. Default is FALSE, meaning that 'dose' is provided on log scale. Set to TRUE if 'dose' is on original scale.}
}
\value{
The row number of the dose combination of interest.
}
\description{
Find the row number of a dose combination of interest.
}

