% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doses.R
\name{doses}
\alias{doses}
\title{Combine original and log doses in a single object}
\usage{
doses(dose_matrix, d_star)
}
\arguments{
\item{dose_matrix}{Doses of interest on original scale.}

\item{d_star}{Reference doses for scaling.}
}
\value{
A list containing the doses of interest on the original and log scale.
}
\description{
Combine original and log doses in a single object
}

