% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_starting_dose.R
\name{find_starting_dose}
\alias{find_starting_dose}
\title{Find the id number of the recommended starting dose}
\usage{
find_starting_dose(doses)
}
\arguments{
\item{doses}{A list containing combination doses.}
}
\value{
Starting dose.
}
\description{
Find the id number of the recommended starting dose
}

