% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_steps.R
\name{first_steps}
\alias{first_steps}
\title{Plot what posterior would look like after first cohort.}
\usage{
first_steps(doses, cohort_size, starting_dose)
}
\arguments{
\item{doses}{List of doses on original and log scales.}

\item{cohort_size}{The size of the first cohort.}

\item{starting_dose}{The starting dose on the original scale.}
}
\value{
A list of ggplot objects.
}
\description{
Plot what posterior would look like after first cohort.
}

