% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_info}
\alias{get_info}
\title{Find quantiles of p(DLT) and probabilites of toxicity}
\usage{
get_info(dose, s)
}
\arguments{
\item{dose}{A vector: dose combination. Log-transformed, and must match what's used in the JAGS model.}

\item{s}{JAGS output from fitting the model.}
}
\value{
A vector of length 8. The first 2 elements simply return the dose levels. The 3rd -- 5th elements
are the 0.1,0.5 and 0.9 quantiles of the distribution of p(DLT). The 6th -- 8th elements are the probabilities
of underdosing, target-dosing and overdosing, respectively.
}
\description{
Provided with a matrix of dose combinations and JAGS output from model, this function returns the
quantiles of p(DLT), the probability that p(DLT) < 0.2, the probability that p(DLT) in (0.2,0.35);
and the  probability that p(DLT) > 0.35.
}

