% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_next_dose.R
\name{get_next_dose}
\alias{get_next_dose}
\title{Get the next dose}
\usage{
get_next_dose(info_mat, doses, current_dose)
}
\arguments{
\item{info_mat}{Matrix containing information about underdosing and overdosing at each dose combination}

\item{doses}{A list containing combination doses.}

\item{current_dose}{Row number of 'doses' corresponding ot current dose.}
}
\value{
The row number corresponding to the dose combination with the largest probability of target
toxicity out of all admissible doses. If all dose combinations are inadmissible then NA is returned.
}
\description{
This function takes into account the probability of underdosing and overdosing at each dose combination
to choose the dose for the next cohort.
}

