% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prior_info.R
\name{get_prior_info}
\alias{get_prior_info}
\title{Find quantiles of p(DLT) and probabilites of toxicity}
\usage{
get_prior_info(dose, prior, n_sim = 1000, seed = 1987)
}
\arguments{
\item{dose}{A vector:  dose combination. Log-transformed doses: must match what's used in the JAGS model.}

\item{prior}{A list of prior parameters}

\item{n_sim}{The number of times to simulate from the prior}

\item{seed}{Random number seed.}
}
\value{
A vector of length 8. The first 2 elements simply return the dose levels. The 3rd -- 5th elements
are the 0.1,0.5 and 0.9 quantiles of the distribution of p(DLT). The 6th -- 8th elements are the probabilities
of underdosing, target-dosing and overdosing, respectively.
}
\description{
Provided with a matrix of dose combinations and the prior, this function returns the
quantiles of p(DLT), the probability that p(DLT) < 0.2, the probability that p(DLT) in (0.2,0.35);
and the  probability that p(DLT) > 0.35.
}

