% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_samples.R
\name{get_samples}
\alias{get_samples}
\title{Use JAGS to get simulations from posterior}
\usage{
get_samples(df, prior, n_iter = 2500)
}
\arguments{
\item{df}{Data frame containing the data}

\item{prior}{A list of prior parameters.}

\item{n_iter}{Number of iterations in MCMC.}

\item{d_star}{A vector of reference levels. Length 2.}
}
\value{
A list, containing samples from the posterior distribution.
}
\description{
Given data and a prior distribution, this function gives back samples from the posterior distribution.
}

