% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_combo_dose_id.R
\name{plot_combo_dose_id}
\alias{plot_combo_dose_id}
\title{Display the id number of each dose combination.}
\usage{
plot_combo_dose_id(doses)
}
\arguments{
\item{doses}{List of doses on original and log scales.}
}
\value{
A ggplot object.
}
\description{
Display the id number of each dose combination.
}

