% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_contour.R
\name{plot_contour}
\alias{plot_contour}
\title{Plot contours of p(overdose) and p(target toxicity).}
\usage{
plot_contour(doses, s, n_points = 50)
}
\arguments{
\item{doses}{A list containing doses on the original and log scale.}

\item{s}{MCMC output...from 'get_samples'}

\item{n_points}{The number of points (on each axis) to use in the contour plot.
then this dose will be declared inadmissible.}
}
\value{
A list of three contour plots: for Pr(overdosing), Pr(target toxicity) and 'best guess' at p(DLT).
}
\description{
Plot contours of p(overdose) and p(target toxicity).
}

