% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_info.R
\name{plot_info}
\alias{plot_info}
\title{Plot the prior/posterior distribution for p(DLT) at all dose combinations}
\usage{
plot_info(info, doses, df = NULL)
}
\arguments{
\item{info}{A matrix with a row for each dose combination, where each row is a vector of length 8.  The 1st -- 3rd elements
are the 0.1,0.5 and 0.9 quantiles of the distribution of p(DLT). The 4th -- 6th elements are the probabilities
of underdosing, target-dosing and overdosing, respectively.}

\item{doses}{A list containing doses on the original and log scale.}

\item{df}{The data.}
}
\value{
A ggplot object.
}
\description{
Plot the prior/posterior distribution for p(DLT) at all dose combinations
}

