% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_route_through_doses.R
\name{plot_route_through_doses}
\alias{plot_route_through_doses}
\title{Display the id number of each dose combination with route through doses.}
\usage{
plot_route_through_doses(doses, route_through_doses, start_rule)
}
\arguments{
\item{doses}{List of doses on original and log scales.}

\item{route_through_doses}{A vector showing order to go through doses.}

\item{start_rule}{Index of 'route_through_doses' corresponding to the starting dose.}
}
\value{
A ggplot object.
}
\description{
Display the id number of each dose combination with route through doses.
}

