% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_selected.R
\name{plot_selected}
\alias{plot_selected}
\title{Plot how often each dose combination is selected}
\usage{
plot_selected(doses, route_through_doses, truth, model_sim, rule_sim)
}
\arguments{
\item{doses}{List of doses on original and log scales.}

\item{route_through_doses}{A vector showing order to go through doses.}

\item{truth}{The 'true' probabilities of DLT at the dose combinations}

\item{model_sim}{a summary of the model based simulation -- from 'sum_sims'.}

\item{rule_sim}{a summary of the model based simulation -- an object from 'crmPack'.}
}
\value{
A ggplot object.
}
\description{
Plot how often each dose combination is selected
}

