% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trial.R
\name{plot_trial}
\alias{plot_trial}
\title{Plot the results of a simulated trial}
\usage{
plot_trial(sim, doses)
}
\arguments{
\item{sim}{The results of a simulated trial. A list. Output of "sim_trial"}

\item{doses}{A list containing doses on the original and log scale.}
}
\value{
A ggplot of the trial.
}
\description{
Plot the results of a simulated trial
}

