% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_truth.R
\name{plot_truth}
\alias{plot_truth}
\title{Plot the 'true' p(DLT) at all dose combinations}
\usage{
plot_truth(doses, truth)
}
\arguments{
\item{doses}{List of doses on original and log scales.}

\item{truth}{A vector. The 'true' p(DLT), corresponding to the rows of 'doses'.}
}
\value{
A ggplot object.
}
\description{
Plot the 'true' p(DLT) at all dose combinations
}

