% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_trial.R
\name{sim_trial}
\alias{sim_trial}
\title{Simulate a combination dose finding trial.}
\usage{
sim_trial(doses, truth, max_n_cohort, cohort_size, starting_dose)
}
\arguments{
\item{doses}{A list containing doses on the original and log scale.}

\item{truth}{A vector of 'true' p(DLT)'s, corresponding to the rows of 'log_doses'.}

\item{max_n_cohort}{The maximum number of cohorts.}

\item{cohort_size}{Cohort size.}

\item{starting_dose}{Starting dose.}
}
\value{
A vector containing the number of patients dosed, the number of dlts, and the recommended MTD.
}
\description{
Simulate a combination dose finding trial.
}

