% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_sims.R
\name{sum_sims}
\alias{sum_sims}
\title{Summarize results from simulation}
\usage{
sum_sims(sims, truth)
}
\arguments{
\item{sims}{list coming from 'sim_study'.}

\item{truth}{a vector of the true p(DLT) at each dose combination.}
}
\value{
A list of information.
}
\description{
Summarize results from simulation
}

