% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_doses.R
\name{combine_doses}
\alias{combine_doses}
\title{Choose dose combinations to be investigated, and plot the prior probabilities of toxicity.}
\usage{
combine_doses(doses_1, doses_2, ref_dose_1, ref_dose_2, prior, over_limit)
}
\arguments{
\item{doses_1}{The doses of drug 1 to be tested.}

\item{doses_2}{The doses of drug 2 to be tested.}

\item{ref_dose_1}{The reference dose of drug 1.}

\item{ref_dose_2}{The reference dose of drug 2.}

\item{prior}{A list of prior parameters.}

\item{over_limit}{A number between 0 and 1. If the probability of overdosing is more than 'over_limit'
then this dose will be declared inadmissible.}
}
\value{
A list of 4 items: dose combinations on original scale, dose combinations on log scale, plot of prior distribution, plot of prior distribution including marginals.
}
\description{
Choose dose combinations to be investigated, and plot the prior probabilities of toxicity.
}

