% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next_steps.R
\name{next_steps}
\alias{next_steps}
\title{Plot what posterior would look like after first cohort.}
\usage{
next_steps(doses, doses_1_so_far, doses_2_so_far, n_so_far, n_dlt_so_far,
  last_dose)
}
\arguments{
\item{doses}{List of doses on original and log scales.}

\item{doses_1_so_far}{The doses of drug 1 tried so far. Vector, on original scale.}

\item{doses_2_so_far}{The doses of drug 2 tried so far (matching positions with doses_1_so_far). Vector, on original scale.}

\item{n_so_far}{The number of patients on each dose combination so far (matching positions with doses_1_so_far). Vector.}

\item{n_dlt_so_far}{The number of dlts on each dose combination so far (matching positions with doses_1_so_far). Vector.}

\item{last_dose}{The most recent dose. Vector of length 2 on the original scale.}
}
\value{
A list of length 2. The first element is the recommended next dose. The second element is a plot of posterior.
}
\description{
Plot what posterior would look like after first cohort.
}

