% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_study.R
\name{sim_study}
\alias{sim_study}
\title{Simulate multiple studies and collect results}
\usage{
sim_study(doses, truth, max_n_cohort, cohort_size, n_sims, starting_dose)
}
\arguments{
\item{doses}{A list containing doses on the original and log scale.}

\item{truth}{A vector of 'true' p(DLT)'s, corresponding to the rows of 'log_doses'.}

\item{max_n_cohort}{The maximum number of cohorts.}

\item{cohort_size}{The number of patients per cohort.}

\item{n_sims}{The number of simulations.}
}
\value{
A matrix of results. 3 columns: n, n_dlt, mtd. One row per simulation.
}
\description{
Simulate multiple studies and collect results
}

